Implement omp flush
This may be considered a bit of a dodgy implementation, like there is
probably a lot more to reflect about with memory consistency etc. but
the benchmarks are not bad so far. Probably will try and add a
simple_flush test, but I think writing such a test is hard