WIP: Add original code (#156)
* Get basic intel_nstream to run

* Change intel inputs to be more interesting

* Implement omp flush

This may be considered a bit of a dodgy implementation, like there is
probably a lot more to reflect about with memory consistency etc. but
the benchmarks are not bad so far. Probably will try and add a
simple_flush test, but I think writing such a test is hard

* Add intel_synch_p2p, it works

* Add simple_flush example

It is honestly not a good example, but even the official OpenMP handbook
does not have good examples at all for flush. I think the support is
dropping.