Make barrier unique_ptr
Gets rid of need for destructor and manual memory management, note that
this is not possible for the nextLevel variable which is the other var
with a new in that file, this is because it is shared with the children
that will join before the end of the function. We could always allocate
the latter on the stack if it becomes an issue, but I figured the OpenMP
stack is small enough it doesn't need big classes allocated on it.