Move OMPLevel storage to WASM thread state
This will enable the distributed code to be much less "if distributed
then X else Y" (to avoid things like the THREAD_MODE) in pthread because
we will need more integration with this.
This implies that the WasmModule will be somewhat aware of OMP, or at
least in the thread specific code which makes sense.