Fix race condition for nested OpenMP levels
In order to be able to test OpenMP level, we need each thread to be able
to individually set possibly a different next number of threads than
other threads of its level. I initially implemented this for simplicity
as "you're forced to have the same number of threads as other threads of
your level". However the new structure with the TLS means that this
wasn't possible anymore for pushed number of threads since this value is
reset at every fork, which raises an issue for a fork within a fork.
An example: first fork creates two threads T1, T2. T1 forks and resets
the pushed threads before T2 forks. T2 now enters fork and reads the
reset value.
This could be solved by having an atomic counter on the reset but this
breaks the case of when a fork follows a fork at the same level. so I
decided to implement this feature properly and just have those variables
be on a separate TLS so threads cannot mess with each other.