Distributed OpenMP experiment 1 (#238)
* Start development branch

* Back to not throwing an error

* Implement derived Levels for reduction types

* Disable CI

* Cleanup

* Add multi_pi and fix format bug

Multi_pi triggers the distributed behaviour but most notably I changed
the inner for loop to have the `nowait` clause on it to get rid of the
barrier call. I may need to rebench the previous stuff with that because
of this unecessary overhead

* Rsynch OMP binaries

* Add debug info in scheduler

* More info

* Convert SchedulerOpinion to enum class with a 0

* Convert SchedulerOpinion from enum to enum class for having type
checking and all that goodness
* Map's [] operator is being used (_sigh in cpp_) which defaults to 0
initialisation, since we want this to be the MAYBE type we need to be
explicit about it

* Make nodeId a reference to avoid it changing

* Make it const for good measure

* Set up permissions on filesystem dirs

* More permission tweaks

* Fix typo in folder

* Sets max number of threads per worker to 5

* Make loop as big as possible in multi_pi

* Implement horrible time hack

* Add debug flags to inke impl

* Add arguments info

* Polish edges

* Final touches to bench script to make it work

* More tests

* Prepares better distributed experiment

* Always record numbers in fork

* Increase number of workers per host

* Limit OMP signle host max by MAX_THREAD_PER_WORKER

* Reduce list range

* Makes function slower

* Add LLVM include folders

This notably helps IDE detection on a system without libgomp

* Use i64 in multi_pi loop

* Implement i64 OMP for loops

* Improve experiments based on local experiences

* Increase the huge loop size to be in the i64 range
* Decrease available number of cores for the worker to be one below the
actual koala limits
* Add backoff to number of times ran in low numbers

* Fix backoff logic

* Add lpush and rpush test

It doesn't check behaviour really, but I won't do strtol on LINDEX
result just yet

* Prepares for review

* Polish up angles

* Re-enable all omp tests

Disabled because of the Clang-10 issue, not the case anymore

* Fix race condition for nested OpenMP levels

In order to be able to test OpenMP level, we need each thread to be able
to individually set possibly a different next number of threads than
other threads of its level. I initially implemented this for simplicity
as "you're forced to have the same number of threads as other threads of
your level". However the new structure with the TLS means that this
wasn't possible anymore for pushed number of threads since this value is
reset at every fork, which raises an issue for a fork within a fork.
An example: first fork creates two threads T1, T2. T1 forks and resets
the pushed threads before T2 forks. T2 now enters fork and reads the
reset value.
This could be solved by having an atomic counter on the reset but this
breaks the case of when a fork follows a fork at the same level. so I
decided to implement this feature properly and just have those variables
be on a separate TLS so threads cannot mess with each other.

Co-authored-by: UG Student <mfournial@koala10.doc.res.ic.ac.uk>